requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
kucoin-python==1.0.11
requests==2.27.1
urllib3==1.26.9
websockets==10.3
